backend/models/scripts/requirements_yolo.txt
ultralytics
torch
torchvision
opencv-python
pillow
pyyaml
numpy
matplotlib
seaborn
pandas
tqdm
